# Cut time to build name lookup index in half