## Overview

* This runs as the last step on every push, since this index is used to serve the API requests on share
* The current version of the code has to load the entire root namespace (including metadata)
* New version uses the lighter-weight V2 branches where we can save some time by only loading the namespace structure itself.

## Implementation notes

Write a new `toNames` for the V2 causal branch, use that instead of the V1 `toNames` which runs on the entire root branch.

I recorded several sample times for how long to completely rebuild the index for the entire 'base' codebase (including all releases); note that these were unoptimized builds:

- Old: 6.87s, 6.98s, 6.57s
- New: 3.69s, 3.51s, 3.38s

So the new one takes about half as long on average.

I can test with optimized builds too, but it seems pretty clear that the new version is faster.

## Test coverage

- [x] I built indexes using the current and new implementations and dumped the contents of the sqlite tables, they were identical.

## Loose ends

There's a lot of unnecessary work going on even in this version, specifically we build a lot of maps and sets only to tear them down, convert their types, then re-build them again. I'll probably investigate a version which skips some of these structures and see if it's faster, but it's possible that we're fusing these structures away anyways 🤷🏼